refactor(api): split project protocol mapping
