+++ pkg/agentcompose/api/project_validation_mapper.go
@@ -1,0 +1,32 @@
+package api
+
+import (
+	"errors"
+	"strings"
+
+	"agent-compose/pkg/compose"
+	agentcomposev2 "agent-compose/proto/agentcompose/v2"
+)
+
+func IssueFromComposeError(err error) *agentcomposev2.ProjectValidationIssue {
+	var validationErr *compose.ValidationError
+	if errors.As(err, &validationErr) {
+		return ProjectValidationIssue(validationErr.Path, validationErr.Message)
+	}
+	var parseErr *compose.ParseError
+	if errors.As(err, &parseErr) {
+		return ProjectValidationIssue(parseErr.Path, parseErr.Message)
+	}
+	return ProjectValidationIssue("spec", err.Error())
+}
+
+func ProjectValidationIssue(path, message string) *agentcomposev2.ProjectValidationIssue {
+	if strings.TrimSpace(path) == "" {
+		path = "spec"
+	}
+	return &agentcomposev2.ProjectValidationIssue{
+		Severity: agentcomposev2.ProjectValidationSeverity_PROJECT_VALIDATION_SEVERITY_ERROR,
+		Path:     path,
+		Message:  message,
+	}
+}
